Rename header files and correct format.
