+++ smacc_sm_reference_library/sm_branching/include/sm_branching/states/st_state_1.hpp
@@ -12,29 +12,37 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <smacc/smacc.h>
+#pragma once
+
+#include <smacc/smacc.hpp>
 
 namespace sm_branching
 {
+using namespace cl_ros_timer;
+using namespace smacc::default_transition_tags;
+
 // STATE DECLARATION
-struct State6 : smacc::SmaccState<State6, SmBranching>
+struct State1 : smacc::SmaccState<State1, SmBranching>
 {
   using SmaccState::SmaccState;
 
   // TRANSITION TABLE
   typedef mpl::list<
 
-    >
+    Transition<EvTimer<CbTimerCountdownOnce, OrTimer>, State2, SUCCESS>,
+    Transition<EvTimer<CbTimerCountdownLoop, OrTimer>, State2b, SUCCESS>,
+    Transition<EvTimer<CbTimerCountdownLoop, OrTimer>, State2c, SUCCESS> >
     reactions;
 
   // STATE FUNCTIONS
   static void staticConfigure()
   {
+    configure_orthogonal<OrTimer, CbTimerCountdownLoop>(3);  // EvTimer triggers each 3 client ticks
     configure_orthogonal<OrTimer, CbTimerCountdownOnce>(
       5);  // EvTimer triggers once at 10 client ticks
   }
 
-  void runtimeConfigure() { RCLCPP_INFO(getLogger(), "Entering State6"); }
+  void runtimeConfigure() {}
 
   void onEntry() { RCLCPP_INFO(getLogger(), "On Entry!"); }
