+++ smacc_sm_reference_library/sm_branching/include/sm_branching/states/st_state_2.hpp
@@ -12,19 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <smacc/smacc.h>
+#pragma once
+
+#include <smacc/smacc.hpp>
 
 namespace sm_branching
 {
 // STATE DECLARATION
-struct State5b : smacc::SmaccState<State5b, SmBranching>
+struct State2 : smacc::SmaccState<State2, SmBranching>
 {
   using SmaccState::SmaccState;
 
   // TRANSITION TABLE
   typedef mpl::list<
 
-    Transition<EvTimer<CbTimerCountdownOnce, OrTimer>, State6, SUCCESS>
+    Transition<EvTimer<CbTimerCountdownOnce, OrTimer>, State3, SUCCESS>
 
     >
     reactions;
@@ -36,7 +38,7 @@ struct State5b : smacc::SmaccState<State5b, SmBranching>
       5);  // EvTimer triggers once at 10 client ticks
   }
 
-  void runtimeConfigure() { RCLCPP_INFO(getLogger(), "Entering State5"); }
+  void runtimeConfigure() { RCLCPP_INFO(getLogger(), "Entering State2"); }
 
   void onEntry() { RCLCPP_INFO(getLogger(), "On Entry!"); }
