+++ smacc_sm_reference_library/sm_branching/include/sm_branching/states/st_state_3c.hpp
@@ -12,19 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <smacc/smacc.h>
+#pragma once
+
+#include <smacc/smacc.hpp>
 
 namespace sm_branching
 {
 // STATE DECLARATION
-struct State2c : smacc::SmaccState<State2c, SmBranching>
+struct State3c : smacc::SmaccState<State3c, SmBranching>
 {
   using SmaccState::SmaccState;
 
   // TRANSITION TABLE
   typedef mpl::list<
 
-    Transition<EvTimer<CbTimerCountdownOnce, OrTimer>, State3c, SUCCESS>
+    Transition<EvTimer<CbTimerCountdownOnce, OrTimer>, State4c, SUCCESS>
 
     >
     reactions;
@@ -36,7 +38,7 @@ struct State2c : smacc::SmaccState<State2c, SmBranching>
       5);  // EvTimer triggers once at 10 client ticks
   }
 
-  void runtimeConfigure() { RCLCPP_INFO(getLogger(), "Entering State2"); }
+  void runtimeConfigure() { RCLCPP_INFO(getLogger(), "Entering State3"); }
 
   void onEntry() { RCLCPP_INFO(getLogger(), "On Entry!"); }
