+++ smacc_sm_reference_library/sm_branching/include/sm_branching/states/st_state_5b.hpp
@@ -12,35 +12,33 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <smacc/smacc.h>
+#pragma once
+
+#include <smacc/smacc.hpp>
 
 namespace sm_branching
 {
-using namespace cl_ros_timer;
-using namespace smacc::default_transition_tags;
-
 // STATE DECLARATION
-struct State1 : smacc::SmaccState<State1, SmBranching>
+struct State5b : smacc::SmaccState<State5b, SmBranching>
 {
   using SmaccState::SmaccState;
 
   // TRANSITION TABLE
   typedef mpl::list<
 
-    Transition<EvTimer<CbTimerCountdownOnce, OrTimer>, State2, SUCCESS>,
-    Transition<EvTimer<CbTimerCountdownLoop, OrTimer>, State2b, SUCCESS>,
-    Transition<EvTimer<CbTimerCountdownLoop, OrTimer>, State2c, SUCCESS> >
+    Transition<EvTimer<CbTimerCountdownOnce, OrTimer>, State6, SUCCESS>
+
+    >
     reactions;
 
   // STATE FUNCTIONS
   static void staticConfigure()
   {
-    configure_orthogonal<OrTimer, CbTimerCountdownLoop>(3);  // EvTimer triggers each 3 client ticks
     configure_orthogonal<OrTimer, CbTimerCountdownOnce>(
       5);  // EvTimer triggers once at 10 client ticks
   }
 
-  void runtimeConfigure() {}
+  void runtimeConfigure() { RCLCPP_INFO(getLogger(), "Entering State5"); }
 
   void onEntry() { RCLCPP_INFO(getLogger(), "On Entry!"); }
