+++ smacc_sm_reference_library/sm_branching/include/sm_branching/states/st_state_6.hpp
@@ -12,19 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <smacc/smacc.h>
+#pragma once
+
+#include <smacc/smacc.hpp>
 
 namespace sm_branching
 {
 // STATE DECLARATION
-struct State3 : smacc::SmaccState<State3, SmBranching>
+struct State6 : smacc::SmaccState<State6, SmBranching>
 {
   using SmaccState::SmaccState;
 
   // TRANSITION TABLE
   typedef mpl::list<
-
-    Transition<EvTimer<CbTimerCountdownOnce, OrTimer>, State4, SUCCESS>
 
     >
     reactions;
@@ -36,7 +36,7 @@ struct State3 : smacc::SmaccState<State3, SmBranching>
       5);  // EvTimer triggers once at 10 client ticks
   }
 
-  void runtimeConfigure() { RCLCPP_INFO(getLogger(), "Entering State3"); }
+  void runtimeConfigure() { RCLCPP_INFO(getLogger(), "Entering State6"); }
 
   void onEntry() { RCLCPP_INFO(getLogger(), "On Entry!"); }
