+++ smacc_sm_reference_library/sm_branching/src/sm_branching/sm_branching_node.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <sm_branching/sm_branching.h>
+#include <sm_branching/sm_branching.hpp>
 
 //--------------------------------------------------------------------
 int main(int argc, char ** argv)
